requirements/base.txt
Pillow==6.2.0
django-environ==0.4.5
django-filter==2.2.0
django-rq==2.1.0
django==2.2.6
djangorestframework==3.10.3
psycopg2-binary==2.8.3
pytz==2019.3
pyyaml==5.1.2
